create a mixins module for possibly shared classes and methods
